disconnected modal added
